+++ src/Adapter/StreamAdapter.php
@@ -7,8 +7,11 @@ use GuzzleHttp\Exception\AdapterException;
 use GuzzleHttp\Exception\RequestException;
 use GuzzleHttp\Message\MessageFactoryInterface;
 use GuzzleHttp\Message\RequestInterface;
+use GuzzleHttp\Stream\InflateStream;
 use GuzzleHttp\Stream\Stream;
 use GuzzleHttp\Stream\LazyOpenStream;
+use GuzzleHttp\Stream\StreamInterface;
+use GuzzleHttp\Stream\Utils;
 
 /**
  * HTTP adapter that uses PHP's HTTP stream wrapper.
@@ -56,47 +59,23 @@ class StreamAdapter implements AdapterInterface
     {
         $request = $transaction->getRequest();
         $stream = $this->createStream($request, $http_response_header);
-
-        if (!$request->getConfig()['stream']) {
-            $stream = $this->getSaveToBody($request, $stream);
-        }
-
-        // Track the response headers of the request
-        $this->createResponseObject($http_response_header, $transaction, $stream);
-    }
-
-    /**
-     * Drain the steam into the destination stream
-     */
-    private function getSaveToBody(RequestInterface $request, $stream)
-    {
-        if ($saveTo = $request->getConfig()['save_to']) {
-            // Stream the response into the destination stream
-            $saveTo = is_string($saveTo)
-                ? new LazyOpenStream($saveTo, 'r+')
-                : Stream::factory($saveTo);
-        } else {
-            // Stream into the default temp stream
-            $saveTo = Stream::factory();
-        }
-
-        while (!feof($stream)) {
-            $saveTo->write(fread($stream, 8096));
-        }
-
-        fclose($stream);
-        $saveTo->seek(0);
-
-        return $saveTo;
+        $this->createResponseObject(
+            $request,
+            $http_response_header,
+            $transaction,
+            new Stream($stream)
+        );
     }
 
     private function createResponseObject(
+        RequestInterface $request,
         array $headers,
         TransactionInterface $transaction,
-        $stream
+        StreamInterface $stream
     ) {
         $parts = explode(' ', array_shift($headers), 3);
         $options = ['protocol_version' => substr($parts[0], -3)];
+
         if (isset($parts[2])) {
             $options['reason_phrase'] = $parts[2];
         }
@@ -104,14 +83,54 @@ class StreamAdapter implements AdapterInterface
         $response = $this->messageFactory->createResponse(
             $parts[1],
             $this->headersFromLines($headers),
-            $stream,
+            null,
             $options
         );
 
+        // Automatically decode responses when instructed.
+        if ($request->getConfig()->get('decode_content')) {
+            switch ($response->getHeader('Content-Encoding')) {
+                case 'gzip':
+                case 'deflate':
+                    $stream = new InflateStream($stream);
+                    break;
+            }
+        }
+
+        // Drain the stream immediately if 'stream' was not enabled.
+        if (!$request->getConfig()['stream']) {
+            $stream = $this->getSaveToBody($request, $stream);
+        }
+
+        $response->setBody($stream);
         $transaction->setResponse($response);
         RequestEvents::emitHeaders($transaction);
 
         return $response;
+    }
+
+    /**
+     * Drain the stream into the destination stream
+     */
+    private function getSaveToBody(
+        RequestInterface $request,
+        StreamInterface $stream
+    ) {
+        if ($saveTo = $request->getConfig()['save_to']) {
+            // Stream the response into the destination stream
+            $saveTo = is_string($saveTo)
+                ? new Stream(Utils::open($saveTo, 'r+'))
+                : Stream::factory($saveTo);
+        } else {
+            // Stream into the default temp stream
+            $saveTo = Stream::factory();
+        }
+
+        Utils::copyToStream($stream, $saveTo);
+        $saveTo->seek(0);
+        $stream->close();
+
+        return $saveTo;
     }
 
     private function headersFromLines(array $lines)
@@ -333,10 +352,6 @@ class StreamAdapter implements AdapterInterface
         &$http_response_header
     ) {
         $url = $request->getUrl();
-        // Add automatic gzip decompression
-        if (strpos($request->getHeader('Accept-Encoding'), 'gzip') !== false) {
-            $url = 'compress.zlib://' . $url;
-        }
 
         return $this->createResource(
             function () use ($url, &$http_response_header, $context) {