+++ tests/server.js
@@ -104,7 +104,7 @@ var GuzzleServer = function(port, log) {
         } else {
             var response = that.responses.shift();
             res.writeHead(response.statusCode, response.reasonPhrase, response.headers);
-            res.end(response.body);
+            res.end(new Buffer(response.body, 'base64'));
             that.requests.push(request);
         }
     };